Expose variable pools in libhegel and narrow pool ids to i64
Variable pools were the only DataSource primitive the C bindings did not
expose, leaving libhegel consumers unable to write stateful tests
(the building block behind `stateful::Variables` / `#[hegel::state_machine]`).

Add `hegel_new_pool`, `hegel_pool_add`, and `hegel_pool_generate`, mirroring
the existing collection primitives, plus a `pool.c` example and smoke-test
coverage.

While wiring this up, narrow the pool/variable id type on the public
`backend::DataSource` trait (and its native/server impls, `NativeVariables`,
and `stateful::Variables`) from `i128` to `i64`. The `i128` was only an
artifact of ciborium's infallible CBOR-integer conversion; pool ids are small
counters and this matches the `i64` already used for collection ids, so the C
API can use `int64_t` losslessly rather than casting.

+++ src/backend.rs
@@ -59,14 +59,14 @@ pub trait DataSource: Send + Sync {
     ) -> Result<(), DataSourceError>;
 
     /// Create a new variable pool. Returns an opaque pool id.
-    fn new_pool(&self) -> Result<i128, DataSourceError>;
+    fn new_pool(&self) -> Result<i64, DataSourceError>;
 
     /// Register a new variable in the pool. Returns the variable id.
-    fn pool_add(&self, pool_id: i128) -> Result<i128, DataSourceError>;
+    fn pool_add(&self, pool_id: i64) -> Result<i64, DataSourceError>;
 
     /// Draw a variable id from the pool.
     /// If `consume` is true, the variable is removed from the pool.
-    fn pool_generate(&self, pool_id: i128, consume: bool) -> Result<i128, DataSourceError>;
+    fn pool_generate(&self, pool_id: i64, consume: bool) -> Result<i64, DataSourceError>;
 
     /// Record a targeting observation for the current test case.
     ///
